# Conflicts:
#	package.json
#	sys/flexyadmin/models/data/Data_Core.php


Former-commit-id: 731b86d6c0b8defdcd099f780119ded004b5050d
Former-commit-id: d943cc9bafb4499fc0ec570dea87494851959b4b [formerly b2e62f6d2b013ea52096c4f4764c7597a5f6d503]
Former-commit-id: fea70d77cf107056c3b3af5b7b14647ea0776072

+++ sys/flexyadmin/helpers/debug_helper.php
@@ -178,18 +178,27 @@ function trace_($a=NULL,$echo=true,$backtraceOffset=1,$max=50,$class='_trace') {
 function trace_sql($sql) {
   $sql = nice_sql($sql);
   echo "<style>._trace,.xdebug-var-dump {position:relative;box-sizing:border-box;width:99%;margin:5px .5%;padding:5px 10px;overflow:auto;color:#000;font-family:courier,serif;font-size:10px;line-height:14px;border:solid 1px #696;border-radius:5px;background-color:#DEA;opacity:.9;z-index:99999;} ._trace pre {font-size:10px;border:none;background:transparent;margin:0;padding:2px;}</style>";
-  echo "<pre contenteditable=\"true\" class=\"_trace\">".highlight_code($sql)."</pre>";
+  if (function_exists('highlight_code'))
+    echo "<pre contenteditable=\"true\" class=\"_trace\">".highlight_code($sql)."</pre>";
+  else 
+    echo "<pre contenteditable=\"true\" class=\"_trace\">".$sql."</pre>";
 }
 
 function nice_sql($sql,$eol="\n") {
   $sql = preg_replace("/(SELECT)\s/uis", "$1$eol", $sql,1);
   $sql = str_replace("`, ", "`, $eol", $sql);
   $sql = preg_replace("/(FROM)\s/uis", "$eol$eol$1 ", $sql,1);
-  $sql = substr_replace( $sql,$eol."ORDER",strrpos($sql,'ORDER'),'5');
   $sql = preg_replace("/(WHERE)\s/uis", $eol.$eol."$1 ", $sql,1);
   $sql = preg_replace("/(SET|LEFT|RIGHT|GROUP)\s/uis", $eol.$eol."$1 ", $sql);
+  $sql = preg_replace("/(\sOR)\s/uis", $eol."$1 ", $sql);
+  $sql = str_replace(" (", "$eol(", $sql);
   return $sql;
 }
+
+function is_sql($txt) {
+  return (preg_match('/^\bSELECT|INSERT|UPDATE|SHOW\b/u', $txt) >=1);
+}
+
 
 /**
  * Geeft een trace van een string-waarde
@@ -303,8 +312,8 @@ function print_ar($array,$return=false,$tabs=0,$brackets="()") {
  * @return string
  * @author Jan den Besten
  */
-function tabs($t,$tab=" ") {
-  // if (IS_AJAX) $tab="\t";
+function tabs($t=0,$tab=" ") {
+  if ($t<0) $t=0;
 	return str_repeat($tab,$t);
 }
